% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepseg.R
\name{DirichReg}
\alias{DirichReg}
\title{Fitting a Dirichlet Regression}
\usage{
DirichReg(
  formula,
  data,
  model = c("common", "alternative"),
  subset,
  sub.comp,
  base,
  weights,
  control,
  verbosity = 0
)
}
\arguments{
\item{formula}{The formula specifying the regression model}

\item{data}{The data set from which the model is constructed}

\item{model}{The parameterization of the model}

\item{subset}{Not used}

\item{sub.comp}{Not used}

\item{base}{Not used}

\item{weights}{Not used}

\item{control}{A list of variables to control the convergence process}

\item{verbosity}{Not used}
}
\value{
A Dirichlet regression model as specified by
\code{\link[DirichletReg]{DirichReg}}
}
\description{
This is a modified form of the
\code{\link[DirichletReg]{DirichReg}} function, adjusted to avoid "model
frame and formula mismatch in model.matrix()" errors that frequently result
when the formula is long and the original function is called from within
another function.
}
