% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepseg.R
\name{get_p_from_model}
\alias{get_p_from_model}
\title{Retrieve Smallest p-value}
\usage{
get_p_from_model(
  model,
  ord,
  breakpoints,
  i,
  allow_interactions,
  ivs,
  parameterization
)
}
\arguments{
\item{model}{The regression model outputted from \link{DirichReg}}

\item{ord}{A non-negative numeric atomic vector indicating the maximum
exponent that was applied to coefficients added to the regression model}

\item{breakpoints}{A vector of all breakpoints used in \code{model}}

\item{i}{The numeric index of the breakpoint for which to obtain the
\emph{t}-statistic}

\item{allow_interactions}{A boolean value indicating whether interactions
were created between independent variables}

\item{ivs}{A data.frame representing values of the independent variables}

\item{parameterization}{The specific parameterization used to create
\code{model}}
}
\value{
A numeric value representing the smallest \emph{p}-value
corresponding to the specified breakpoint
}
\description{
This function obtains the smallest \emph{p}-value among all
\emph{p}-value for a given breakpoint in a regression model prepared by
prepared by \code{\link[Compositional]{comp.reg}}.
}
