% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepseg.R
\name{get_t_from_model}
\alias{get_t_from_model}
\title{Retrieve Largest t-statistic}
\usage{
get_t_from_model(model, ord, i, allow_interactions, ivs)
}
\arguments{
\item{model}{The regression model outputted from
\code{\link[Compositional]{comp.reg}}}

\item{ord}{A non-negative numeric atomic vector indicating the maximum
exponent that was applied to coefficients added to the regression model}

\item{i}{The numeric index of the breakpoint for which to obtain the
\emph{t}-statistic}

\item{allow_interactions}{A boolean value indicating whether interactions
were created between independent variables}

\item{ivs}{A data.frame representing values of the independent variables}
}
\value{
A numeric value representing the largest absolute value among all
\emph{t}-statistics corresponding to the specified breakpoint
}
\description{
This function obtains the largest absolute value among all
\emph{t}-statistics for a given breakpoint in a regression model summary
prepared by \code{\link[Compositional]{comp.reg}}.
}
