% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepseg.R
\name{print.stepseg_output}
\alias{print.stepseg_output}
\title{Print}
\usage{
\method{print}{stepseg_output}(x, ...)
}
\arguments{
\item{x}{An object with the \code{stepseg_output} class}

\item{...}{Other arguments inherited from the generic \code{print} function}
}
\description{
This method prints the \code{output} element of an object with the
\code{stepseg_output} class, such as output obtained from the
\code{\link{stepseg}} function.
}
