% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepseg.R
\name{stepsegdir}
\alias{stepsegdir}
\title{Stepwise Segmented Dirichlet Regression Analysis}
\usage{
stepsegdir(
  dv,
  ivs = data.frame(1:nrow(data.frame(dv))),
  start_breakpoints = numeric(0),
  retain_start_breakpoints = TRUE,
  add = 0.15,
  remove = 0.2,
  bonferroni = TRUE,
  add_mode = "mse",
  ord = 1,
  parameterization = "common",
  allow_interactions = FALSE,
  iterlim = 1000,
  tol1 = 1e-05,
  tol2 = 1e-10,
  update = 0,
  verbose = FALSE
)
}
\arguments{
\item{dv}{A data.frame representing values of the dependent variable}

\item{ivs}{A data.frame representing values of the independent variables,
coerced to a data.frame if not provided as one (default =
\code{data.frame(1:nrow(data.frame(dv)))}, which implicitly treats
\code{dv} as sequentially ordered data and attempts to detect breakpoints
in that sequence)}

\item{start_breakpoints}{A numeric vector of breakpoints to be added to the
regression model before the first iteration of the stepwise procedure}

\item{retain_start_breakpoints}{If \code{TRUE}, then the breakpoints
specified in \code{start_breakpoints} can never be removed from the model}

\item{add}{A numeric atomic vector between \code{0} and \code{1} indicating
the \emph{p}-value threshold to add coefficients to the model during each
forward selection iteration, which must be less than or equal to the
\code{remove} argument in order to avoid infinite loops (default =
\code{.15}, as recommended by Britt, 2015)}

\item{remove}{A numeric atomic vector between \code{0} and \code{1}
indicating the \emph{p}-value threshold to remove coefficients from the
model during each backward selection iteration, which must be greater than
or equal to the \code{add} argument in order to avoid infinite loops
(default = \code{.20}, as recommended by Britt, 2015)}

\item{bonferroni}{If \code{TRUE}, then the \emph{p}-values specified in
\code{add} and \code{remove} are divided by the number of terms added to
the model for each breakpoint, including all categories of the dependent
variable, all independent variables (including any interactions), and all
orders of those independent variables (\code{0:ord})}

\item{add_mode}{A character atomic vector (either \code{"p"} or \code{"mse"})
indicating what criterion should be used to select the best candidate
block during each forward selection iteration (default = \code{"mse"})}

\item{ord}{A non-negative numeric atomic vector indicating the maximum
exponent that will be applied to coefficients added to the regression
model (default = \code{1}; it is generally recommended to use either
\code{order = 0} or \code{order = 1})}

\item{parameterization}{The parameterization used to perform Dirichlet
regression; since \link{DirichReg} does not return standard errors for
\code{"lmfit"} or \code{"spatial"}, at present, this
must be set to \code{"classical"}}

\item{allow_interactions}{A boolean value indicating whether interactions
may be created between indicator functions representing breakpoints along
one independent variable with other independent variables in the model
(default = \code{FALSE}, which restricts each indicator function to only
interact with the independent variable along which its breakpoint was
defined)}

\item{iterlim}{The maximum number of iterations permitted for convergence in
each regression, as conducted by \link{DirichReg}}

\item{tol1}{The convergence criterion for BFGS optimization in each
regression, as conducted by \link{DirichReg}}

\item{tol2}{The convergence criterion for NR optimization in each regression,
as conducted by \link{DirichReg}}

\item{update}{A numeric value indicating how many loop iterations should
elapse between progress updates (default = \code{0}, which suppresses
output)}

\item{verbose}{A boolean value indicating whether the current regression
model and results should be outputted as part of the progress report after
every step of the stepwise procedure (default = \code{FALSE}}
}
\value{
The final Dirichlet regression model outputted from \link{DirichReg}
}
\description{
This function performs stepwise segmented Dirichlet regression
analysis, a modified form of the procedure described by Britt (2015).
}
\details{
When using any form of stepwise segmented regression analysis,
coefficients are added to and removed from the model in "blocks," with a
given block consisting of the indicator function representing a given
breakpoint location as well as all interaction terms between that indicator
function and independent variables that are eligible to be added to the
model.

The \code{ord} and \code{allow_interactions} arguments jointly indicate what
interaction terms are valid. \code{ord} indicates the maximum exponent that
can be applied to the independent variables, while \code{allow_interactions}
indicates whether interaction terms can be created between a given
independent variable and an indicator function that is defined by values of a
different independent variable.

All interaction terms that are considered valid, with exponents ranging from
\code{0} to \code{ord}, are included in a given block. Consider, for
instance, an analysis in which \code{ivs}, the data.frame representing the
independent variables, has three columns signifying three variables. If
\code{ord = 1} and \code{allow_interactions = FALSE}, which are their default
values in \code{stepseg}, then the block corresponding to a potential
breakpoint at \code{ivs[, 1] = 10} would include two coefficients:
\itemize{
\item{\code{I(ivs[, 1] = 10)}}
\item{\code{I(ivs[, 1] = 10) * (ivs[, 1]^1)}}
}
Setting \code{ord = 2} would retain both of the aforementioned
coefficients and also add
\itemize{
\item{\code{I(ivs[, 1] = 10) * (ivs[, 1]^2)}}
}
to the block. Maintaining \code{ord = 2} and setting
\code{allow_interactions = TRUE} would result in the block containing a
total of seven coefficients:
\itemize{
\item{\code{ivs[, 1] = 10}}
\item{\code{I(ivs[, 1] = 10) * (ivs[, 1]^1)}}
\item{\code{I(ivs[, 1] = 10) * (ivs[, 1]^2)}}
\item{\code{I(ivs[, 1] = 10) * (ivs[, 2]^1)}}
\item{\code{I(ivs[, 1] = 10) * (ivs[, 2]^2)}}
\item{\code{I(ivs[, 1] = 10) * (ivs[, 3]^1)}}
\item{\code{I(ivs[, 1] = 10) * (ivs[, 3]^2)}}
}
Note that if \code{ord = 0}, then each block will only include the
indicator function itself (in this example, \code{I(ivs[, 1] = 10)})
regardless of the value of \code{allow_interactions}.

During each forward selection iteration, the algorithm selects the "best"
block of coefficients that can be added. If \code{add_mode = "p"}, then all
possible candidate blocks that could be added to the model are evaluated,
and whichever block has a \emph{p}-value less than or equal to all
\emph{p}-values in all other blocks is treated as the "best" block. As long
as that \emph{p}-value is less than the threshold specified by the
\code{add} argument, all coefficients included in the block are added to the
model. If \code{add_mode = "mse"}, then the candidate block whose
coefficients would jointly reduce the model mean squared error by the
greatest amount is treated as the \code{best} block, and all of its coefficients
are added to the model if the resulting \emph{p}-value for at least one of
those coefficients would be less than \code{add}.

During each backward selection iteration, all coefficients listed in
\code{start_formula} are retained in the regression model. Among all
remaining coefficients, any blocks whose \emph{p}-values are all greater
than or equal to \code{remove} are removed from the model. If at least one
\emph{p}-value contained in a given block is less than \code{remove}, then
none of the coefficients in the block are removed from the model.

When \code{ord = 0}, \code{add_mode = "p"} is generally acceptable to follow
common conventions of stepwise model selection. When \code{ord > 0}, however,
\code{add_mode = "p"} is more likely to result in spurious breakpoints being
added to the model due to intercept and higher-order terms competing with one
another, and in rare cases the algorithm may entirely fail to converge.
\code{add_mode = "mse"} is more robust against these issues and is generally
recommended as whenever \code{ord > 0}.

As a caveat, since the Dirichlet distribution is multidimensional in nature,
there may be rare cases in which the most suitable breakpoint selected based
on MSE does not correspond to a statistically significant change in any
individual category of the dependent variable. This can cause the model to
prematurely terminate. If an inspection of the final model suggests that such
an issue has occurred, then setting \code{add_mode = "p"} may allow that
problem to be overcome, regardless of other weaknesses in that option.

The larger the value of \code{ord}, the more likely spurious breakpoints are
to emerge. As with other regression contexts, you should only increase the
complexity of your model (such as by increasing the value of \code{ord}) when
you have a clear reason to do so. Moreover, whenever \code{ord > 0},
singularities may occur that render some coefficients inestimable. For
instance, if \code{ord = 3} and a pair of breakpoints is identified along the
same independent variable, with those two breakpoints occurring two data
points apart, there will be insufficient data between the breakpoints to
estimate the standard error of the interaction between the indicator function
and a cubic term, so it will be reported as \code{NA}. This is normal,
expected behavior in stepwise segmented regression analysis. In such cases,
those \code{NA} coefficients can be treated as though they were absent from
the model.

Finally, use caution when setting \code{ord > 1}, as increasingly high-order
terms may compete with lower-order terms in the model and yield unpredictable
results. Refer to Britt (2015) for more information.
}
\section{References}{

Britt, B. C. (2015). Stepwise segmented regression analysis: An iterative
statistical algorithm to detect and quantify evolutionary and
revolutionary transformations in longitudinal data. In S. A. Matei, M. G.
Russell, & E. Bertino (Eds.), \emph{Transparency in social media: Tools,
methods, and algorithms for mediating online interactions} (pp. 125-144).
Heidelberg, Germany: Springer.
}

\examples{
set.seed(2023)
dv <- rbind(DirichletReg::rdirichlet(12, c(8,8,32)),
            DirichletReg::rdirichlet(18, c(40,24,4)))
model <- stepsegdir(dv, add_mode="mse", update=10, verbose=TRUE)
# "Beginning the stepwise procedure..."
# "Testing observation 10"
# "Testing observation 20"
# "End of step 1 ~ Current Breakpoints: 10"
# "Testing observation 10"
# "Testing observation 20"
# "No additional breakpoints can be added to the model."
# Warning message:
# In min(abs(summary(model)$coef.mat[indices, 4]), na.rm = T) :
#   no non-missing arguments to min; returning Inf
summary(model)
#
# Call:
# DirichReg(formula = as.formula(paste0("dv ~ ", paste(current_terms, collapse
# = " + "))), data = df, model = parameterization, control = list(iterlim =
# iterlim, tol1 = tol1, tol2 = tol2))
#
# Standardized Residuals:
#         Min       1Q   Median      3Q     Max
# V1  -2.1452  -0.6651  -0.2261  0.6674  2.1566
# V2  -1.7372  -1.1093   0.3604  0.8497  1.6465
# V3  -1.7572  -0.6547  -0.2742  0.6269  2.2728
#
# ------------------------------------------------------------------
# Beta-Coefficients for variable no. 1: V1
#                                    Estimate Std. Error z value Pr(>|z|)    
# (Intercept)                         4.57845    0.68780   6.657  2.8e-11 ***
# I(ivs[, 1]^1)                      -0.22604    0.09559  -2.365   0.0180 *  
# I(ivs[, 1] > 12)TRUE               -0.54752    1.24725  -0.439   0.6607    
# I(ivs[, 1]^1):I(ivs[, 1] > 12)TRUE  0.21705    0.10660   2.036   0.0417 *  
# ------------------------------------------------------------------
# Beta-Coefficients for variable no. 2: V2
#                                    Estimate Std. Error z value Pr(>|z|)    
# (Intercept)                         4.35351    0.69899   6.228 4.72e-10 ***
# I(ivs[, 1]^1)                      -0.18497    0.09834  -1.881    0.060 .  
# I(ivs[, 1] > 12)TRUE               -0.84702    1.24069  -0.683    0.495    
# I(ivs[, 1]^1):I(ivs[, 1] > 12)TRUE  0.17734    0.10875   1.631    0.103    
# ------------------------------------------------------------------
# Beta-Coefficients for variable no. 3: V3
#                                    Estimate Std. Error z value Pr(>|z|)    
# (Intercept)                         5.97326    0.68569   8.711  < 2e-16 ***
# I(ivs[, 1]^1)                      -0.21326    0.09505  -2.244  0.02485 *  
# I(ivs[, 1] > 12)TRUE               -3.70315    1.21792  -3.041  0.00236 ** 
# I(ivs[, 1]^1):I(ivs[, 1] > 12)TRUE  0.18075    0.10538   1.715  0.08630 .  
# ------------------------------------------------------------------
# Significance codes: 0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#
# Log-likelihood: 118.8 on 12 df (83 BFGS + 2 NR Iterations)
# AIC: -213.5, BIC: -196.7
# Number of Observations: 30
# Link: Log
# Parametrization: common
}
